Refactor ApiResponse.vue component to improve JSON display, add copy to clipboard functionality and refactor Metrics
